+++ app/src/main/java/br/com/concrete/tentacle/features/menu/MenuFragment.kt
@@ -1,5 +1,7 @@
 package br.com.concrete.tentacle.features.menu
 
+import android.content.DialogInterface
+import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -8,22 +10,35 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import br.com.concrete.tentacle.BuildConfig
 import br.com.concrete.tentacle.R
-import br.com.concrete.tentacle.data.models.ErrorResponse
 import br.com.concrete.tentacle.data.models.User
 import br.com.concrete.tentacle.data.models.ViewStateModel
-import kotlinx.android.synthetic.main.fragment_menu.*
+import br.com.concrete.tentacle.data.repositories.SharedPrefRepository
+import br.com.concrete.tentacle.features.login.LoginActivity
+import br.com.concrete.tentacle.utils.DialogUtils
+import br.com.concrete.tentacle.utils.LogWrapper
+import kotlinx.android.synthetic.main.fragment_menu.logout
+import kotlinx.android.synthetic.main.fragment_menu.name
+import kotlinx.android.synthetic.main.fragment_menu.state
+import kotlinx.android.synthetic.main.fragment_menu.version
+import org.koin.android.ext.android.inject
 import org.koin.android.viewmodel.ext.android.viewModel
 
-class MenuFragment : Fragment(), View.OnClickListener {
+class MenuFragment : Fragment() {
 
     private val menuViewModel: MenuViewModel by viewModel()
+    private val sharePrefRepository: SharedPrefRepository by inject()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        initObservable()
         return inflater.inflate(R.layout.fragment_menu, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        init()
+        initObservable()
+        super.onViewCreated(view, savedInstanceState)
     }
 
     private fun initObservable(){
@@ -32,8 +47,8 @@ class MenuFragment : Fragment(), View.OnClickListener {
                 ViewStateModel.Status.SUCCESS -> {
                     base.model?.let { updateUI(it) }
                 }
-                ViewStateModel.Status.ERROR -> {
-                    callError(base.errors)
+                else -> base.errors?.message?.let {
+                    LogWrapper.log("A problem happens", it[0])
                 }
             }
         })
@@ -41,23 +56,37 @@ class MenuFragment : Fragment(), View.OnClickListener {
         menuViewModel.loadUser()
     }
 
-    private fun updateUI(user: User){
+    fun init(){
+        logout.setOnClickListener { checkLogout() }
         version.text = String.format(getString(R.string.version), BuildConfig.VERSION_NAME)
+    }
+
+    private fun updateUI(user: User){
         name.text = user.name
         state.text = user.state.name
     }
 
-    private fun callError(errors: ErrorResponse?) {
-        errors?.let {
-
+    private fun checkLogout() {
+        activity?.let {
+            DialogUtils.showDialog(
+                context = it,
+                title = getString(R.string.logout_title),
+                message = getString(R.string.logout_question),
+                positiveText = getString(R.string.ok),
+                positiveListener = DialogInterface.OnClickListener { _, _ ->
+                    performLogout()
+                },
+                negativeText = getString(R.string.cancel)
+            )
         }
     }
 
-
-    override fun onClick(v: View?) {
-        when(v?.id){
-
-        }
+    private fun performLogout() {
+        sharePrefRepository.removeSession()
+        val login = Intent(activity, LoginActivity::class.java)
+        login.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK)
+        startActivity(login)
+        activity?.finish()
     }
 
 }
