+++ app/src/main/java/br/com/concrete/tentacle/features/menu/MenuViewModel.kt
@@ -6,11 +6,14 @@ import br.com.concrete.tentacle.data.models.ErrorResponse
 import br.com.concrete.tentacle.data.models.User
 import br.com.concrete.tentacle.data.models.ViewStateModel
 import br.com.concrete.tentacle.data.repositories.SharedPrefRepository
+import br.com.concrete.tentacle.data.repositories.UserRepository
 import br.com.concrete.tentacle.utils.Event
+import br.com.concrete.tentacle.utils.LogWrapper
 import br.com.concrete.tentacle.utils.PREFS_KEY_USER
 
 class MenuViewModel(
-    private val sharedPrefRepository: SharedPrefRepository
+    private val sharedPrefRepository: SharedPrefRepository,
+    private val userRepository: UserRepository
 ): BaseViewModel(){
 
     private val stateModel: MutableLiveData<ViewStateModel<User>> = MutableLiveData()
@@ -21,11 +24,18 @@ class MenuViewModel(
         user?.let{
             stateModel.postValue(ViewStateModel(status = ViewStateModel.Status.SUCCESS, model = it))
         } ?: run {
-            stateModel.postValue(ViewStateModel(status = ViewStateModel.Status.SUCCESS, errors = ErrorResponse()))
+            loadUserFromServer()
         }
     }
 
-    fun loadUserFromServer(){
-
+    private fun loadUserFromServer(){
+        disposables.add(
+            userRepository.getProfile().subscribe({
+                sharedPrefRepository.saveUser(PREFS_KEY_USER,it.data)
+                stateModel.postValue(ViewStateModel(status = ViewStateModel.Status.SUCCESS, model = it.data))
+            },{
+                LogWrapper.log("UserProfile: ", it.localizedMessage.toString())
+            })
+        )
     }
 }
